refactor(tools): move tool modules under src/tools
