+++ src/tools/editing.js
@@ -2,8 +2,8 @@ import { z } from "zod";
 import fs from "node:fs";
 import matter from "gray-matter";
 import yaml from "js-yaml";
-import { createSnapshot, listSnapshots } from "../src/core/git.js";
-import { getFileWriteDiagnostics, readMeta, indexSceneFile } from "../src/sync/sync.js";
+import { createSnapshot, listSnapshots } from "../core/git.js";
+import { getFileWriteDiagnostics, readMeta, indexSceneFile } from "../sync/sync.js";
 
 function renderSceneContent(metadata, revisedProse) {
   const hasFrontmatter = metadata && Object.keys(metadata).length > 0;
